Progress Log:
Added the Health System script to the Player and created a Prefab of the Hearts that will be displayed onscreen.
Worked on the Item Script and laid the groundwork for the Consumables System in it.
Attempted to create a Swinging Axe using Physics, failed miserably.
--------------------------------------------------
Issues:

Still need to figure out how to make the Axe swing using Physics, if possible at all.
Item script still needs to be able to detect when the player has run out of a specific item and make it unusable until the player picks up another one.
HighScore System and Rat AI still need to be worked upon.
--------------------------------------------------
Summary:

Most of my week was spent trying to understand the Inventory system and trying to figure out how to recreate the movements of a pendulum using Physics. While I was able to get a basic understanding of the Inventory System, I believe it's still not enough to be able to complete the Consumables part of the Item script. On the other hand, most tutorials on creating a Pendulum in Unity are for Unity2D, which is a problem since modifying code designed for a 2D pendulum to work in a 3D environment is not possible. At least as far as I know.

The Rat AI was delayed mostly due to the fact that we've decided to use Brendan's A* System on nearly all AI entities. I still need to create a script that lets the rats run around, and maybe have them run away from the Golems which the PLayer could use to track if the Golem was moving it their direction or not, but I need to be able to see Brendan's A* System before I do anything. The High Score System was delayed mostly because the other tasks take higher priority at the time.
--------------------------------------------------