Progress Log:
The Health System was converted into a prefab.
The Code for using Pickups was moved from the Item script and integrated into the Inventory Script. (My code has been separated via comments to avoid confusion)
A new scene was used to test the pickup and consumable system.
--------------------------------------------------
Issues:

When the player picks up an item with an amount greater than it's max stack size, that surplus is not moved into the next empty slot, and is instead stored in the same slot.
The sprite for the pickup is not deleted when the player uses up all of the items in a single slot.
--------------------------------------------------
Summary:

The Inventory system is now responsible for sorting all pickups collected into their corresponding lists, but the system isn't as polished as I would like it to be. I'll need to work with Dylan in order to fine tune this system and clear out all the bugs. The Inventory will now check the id of the item being collected and sort it into it's corresponding List. At the same time, it will set the corresponding bool for the item to true, allowing the player to use the items in their inventory until they run out, which is when the bool is set to false.
--------------------------------------------------