Progress Log:
Created a Score System.
Created a UI Element that reflects how many coins the player has collected.
--------------------------------------------------
Issues:

There isn't any known issue with the system in it's current state.
--------------------------------------------------
Summary:

The Score System essentially uses the OnTriggerEnter() to check if the item collected has a tag called "Coin". If true, the scoreCount goes up by one, and this is then reflected in the UI.
--------------------------------------------------